+++ add_trailing_comma/__main__.py
@@ -1,0 +1,4 @@
+from add_trailing_comma._main import main
+
+if __name__ == '__main__':
+    exit(main())
