+++ setup.cfg
@@ -21,14 +21,19 @@ classifiers =
     Programming Language :: Python :: Implementation :: PyPy
 
 [options]
-py_modules = add_trailing_comma
+packages = find:
 install_requires =
     tokenize-rt>=3.0.1
 python_requires = >=3.6.1
 
 [options.entry_points]
 console_scripts =
-    add-trailing-comma = add_trailing_comma:main
+    add-trailing-comma = add_trailing_comma._main:main
+
+[options.packages.find]
+exclude =
+    tests*
+    testing*
 
 [bdist_wheel]
 universal = True
